move statistics to own view
